requirements.txt
arrow==0.12.0
astroid==1.5.3
boto==2.48.0
certifi==2017.11.5
chardet==3.0.4
colorama==0.3.9
dj-database-url==0.4.2
Django==1.11.7
django-forms-bootstrap==3.1.0
django-gravatar==0.1.0
django-gravatar2==1.4.2
django-storages==1.6.5
djangorestframework==3.7.3
gunicorn==19.7.1
idna==2.6
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
Pygments==2.2.0
pylint==1.7.4
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
wrapt==1.10.11
